requirements.txt
bluezero==0.9.0
colorzero==2.0
dbus-python==1.3.2
pycairo==1.27.0
PyGObject==3.50.0
uuid==1.30
